# Conflicts:
#	src/main/java/application/Main.java
#	src/main/java/application/StatsController.java
#	src/main/java/server/Request.java
#	src/main/java/server/queries/VegMealQuery.java

+++ src/main/java/calculator/BusCalculator.java
@@ -14,7 +14,7 @@ public class BusCalculator {
      * @return Passing the values of the distance(int) and kg CO2(double) to the method 'carBus()'.
      */
 
-    public static int bus(int number) {
+    public static Double bus(int number) {
 
         //Setup https client
         String host = "http://impact.brighterplanet.com/";
@@ -50,7 +50,7 @@ public class BusCalculator {
             return busCalculator.carBus(number, co2Bus);
         } catch (MalformedURLException e) {
             e.printStackTrace();
-            return -1;
+            return -1.0;
         }
     }
 
@@ -58,12 +58,10 @@ public class BusCalculator {
      * The calculator for calculating the difference in kg CO2 produced by a bus and a car.
      * @param dist The distance from the method 'bus()'.
      * @param busCo2 The amount of kg CO2 by the bus, from the method 'bus()'.
-     * @return the calculated result of the equation ->
-     * (Amount of kg CO2 produced by a car) - (Amount of kg CO2 produced by a bus/ Average amount of people taking the bus) = kg CO2 saved
-     * @throws MalformedURLException
+     * @return The calculated result of the equation.
+     * @throws MalformedURLException Use of URL
      */
-
-    public static int carBus(int dist, double busCo2) throws MalformedURLException {
+    public static Double carBus(int dist, double busCo2) throws MalformedURLException {
         //Setup https client
         String host = "http://impact.brighterplanet.com/";
         String model = "automobile_trips.json";
@@ -92,7 +90,7 @@ public class BusCalculator {
         //Calculate amount of kg CO2 saved.
         Double result = co2New - (busCo2 / 20);
         result = Math.round(result * 100.00) / 100.00;
-        return result.intValue();
+        return result;
     }
 
 }
